Add modular configuration system and TODO list
- Introduced a modular configuration system for the logging framework, enhancing maintainability and extensibility.
- Added handlers for various configuration keys including `level`, `propagate`, `pipelines`, `custom_levels`, and `async`.
- Implemented validation and application logic for each configuration key to ensure robust configuration management.
- Created a new TODO file to track future enhancements, including verbosity flags handling and live level changing through environment variables.
