+++ .todo.md
@@ -1,0 +1,5 @@
+# TODO
+
+- [ ]  TODO: add verbosity flags handling
+- [ ] TODO: add live level changing through env vars.
+
